tests/wpt/web-platform-tests/encoding-detection/yi-windows-1255-late.tentative.html

<!doctype html>
<meta charset="utf-8">
<title>yi windows-1255 late</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script>
setup({explicit_done:true});
window.onmessage = function(e) {
    test(function() {
        assert_equals(e.data, "windows-1255", 'Expected windows-1255');
    }, "Check detection result");
    w.close();
    done();
};
var w = window.open("support/yi-windows-1255-late.sub.html");
</script>
